Add checksum
